Added a few block/item sets (Corruption, Crimson, Hallow, Chests, Doors) that can be used to highlight/find groups of blocks/items. I'll add more later, and explore the ability for user-created custom sets. Fixed ability to find Shark Statues, Demon Altars, Crimson Altars, Shadow Orbs, Crimson Hearts, and Spiky Ball Traps. Fixed chest item counts overflowing (999 was showing as 231). Fixed chest item popup not scrolling if the list doesn't fit on the screen. Fixed tooltips for some toolbar buttons. Search box is now automatically focused when blocks drop down is opened.

+++ Solution/Data/Item.cs
@@ -7,7 +7,7 @@ namespace TerraMap.Data
 {
 	public class Item
 	{
-		public byte Count { get; set; }
+		public int Count { get; set; }
 		public string Name { get; set; }
 		public int Id { get; set; }
 		public byte PrefixId { get; set; }
